Refactor a folder name
